feat: Update web UI to a modern, responsive design
This commit updates the web interface of the E-Book Librarian to a more modern and responsive design using Vue.js.

The following files were updated:
- main/web_assets/index.html
- main/web_assets/script.js
- main/web_assets/style.css

Additionally, the `main/CMakeLists.txt` was updated to fix a build issue where the `spiffs_create_partition_image` command was called before the `spiffs` component was registered.
