+++ main/web_assets/style.css
@@ -1,0 +1,17 @@
+body { font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif; margin: 0; background-color: #f0f2f5; color: #1c1e21; }
+.container { max-width: 900px; margin: 20px auto; padding: 20px; background-color: #fff; border-radius: 8px; box-shadow: 0 2px 4px rgba(0,0,0,0.1); }
+h1 { color: #1877f2; text-align: center; border-bottom: 1px solid #ddd; padding-bottom: 10px; }
+.status { padding: 10px; background-color: #e7f3ff; border: 1px solid #b7d9f7; border-radius: 6px; margin-bottom: 20px; text-align: center; }
+.status.connected { background-color: #e9f7ef; border-color: #b7e4c7; }
+.status.disconnected { background-color: #fff1f0; border-color: #ffccc7; }
+.panels { display: flex; flex-wrap: wrap; gap: 20px; }
+.panel { flex: 1; min-width: 300px; background: #f9f9f9; padding: 15px; border-radius: 6px; border: 1px solid #e0e0e0; }
+.panel h2 { margin-top: 0; font-size: 1.2em; }
+ul { list-style-type: none; padding: 0; max-height: 400px; overflow-y: auto; }
+li { padding: 8px 10px; border-bottom: 1px solid #eee; display: flex; align-items: center; justify-content: space-between; word-break: break-all; }
+li:last-child { border-bottom: none; }
+.file-name { cursor: pointer; }
+.file-name.selected { font-weight: bold; color: #1877f2; }
+button { background-color: #1877f2; color: white; border: none; padding: 10px 15px; border-radius: 6px; cursor: pointer; font-size: 1em; transition: background-color 0.2s; }
+button:disabled { background-color: #a0c3f7; cursor: not-allowed; }
+.actions { margin-top: 20px; text-align: center; }
