+++ partitions.csv
@@ -1,0 +1,5 @@
+# Name,   Type, SubType, Offset,  Size, Flags
+nvs,      data, nvs,     ,        0x9000,
+otadata,  data, ota,     ,        0x2000,
+app0,     app,  ota_0,   ,        1M,
+storage,  data, spiffs,  ,        1M,
